requirements.txt
streamlit
requests
matplotlib
reportlab
pandas
Pillow
tqdm
folium
streamlit-folium
altair
jinja2
plotly
psycopg[binary]
playwright
